pyproject.toml
[tool.poetry]
name = "finance-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.0"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
pydantic = "^1.10.2"
sqlalchemy = "^1.4.44"
psycopg2-binary = "^2.9.5"
python-jose = "^3.3.0"
cryptography = "^39.0.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.5"
python-dotenv = "^0.21.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
